게시물 수정, 삭제 게시물 존재여부체크, 로그인체크, 권한체크

+++ src/main/webapp/jsp/article/detail.jsp
@@ -5,6 +5,7 @@
     
 <%
 	Map<String, Object> articleMap = (Map<String, Object>) request.getAttribute("articleMap");
+	int loginedMemberId = (int) request.getAttribute("loginedMemberId");
 %>    
 
 <!DOCTYPE html>
@@ -45,8 +46,11 @@
 	
 	<div>
 		<a href="list">목록</a>
-		<a href="modify?id=<%= articleMap.get("id") %>">수정</a>
-		<a href="doDelete?id=<%= articleMap.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+		
+		<% if (loginedMemberId != -1 && (int) articleMap.get("memberId") == loginedMemberId) { %>
+			<a href="modify?id=<%= articleMap.get("id") %>">수정</a>
+			<a href="doDelete?id=<%= articleMap.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+		<% } %>
 	</div>
 </body>
 </html>
